requirements.txt
azure-functions==1.17.0
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
aiohttp>=3.9.5
openai>=1.46.0
